% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survicompet.R
\name{SurviCompet}
\alias{SurviCompet}
\title{SurviCompet}
\usage{
SurviCompet(m, n, data)
}
\arguments{
\item{m}{numeric staring column of variables}

\item{n}{numeric ending column of variables}

\item{data}{Data frame by user}

\item{Status}{Column states 0, 1, 2 on the basis of Prog column and Death column}
}
\description{
'SurviCompet' function from 'progcomprisk' enables the user to obtain DEGs
(Differentially expressed genes) which leads to death given it leads to
progression of cancer as well as accounts for competing risks.
}
\details{
SurviCompet function first adds a new column 'Status' to the input
dataset and assigns status 0, 1, 2 to each cell. It assigns 0 when progression is
1, death is 0 or progression 0 death is 0, assigns 1 when progression is 1 and
death is 1, whereas assigns 2 when progression is 0 and death is 1.

Further, it subsets 2 data sets, one data set named 'deathdata' which includes Ids/
Subjects with status 0 and 1 and applies Cox PH on it. Another subset data named as
'compdata' includes Ids/Subjects with status 0 and 2, then applies Cox PH after
substituting 2 by 1. Then it filters out DEGs having P-value > 0.05 from both
subset data. At last the it merges the commom DEGs from both data and creates a
new dataframe which consists 'Id','OS','Death','PFS','Prog','Status' and observations
of common significant DEGs(leading to death given they leads to progression of cancer
as well as accounts for competing risks)
function it lists out the common DEGs from


SurviCompet(m,n,data), m - starting
column number of data from where DEG observations are to be read,    n - ending
column number of data till which DEG observations are to be read,    data - name of
input data consisting survival and progression outcomes with DEG observations.

Concern should be given to rename the columns of the input data before using 'progcomprisk' package, as follows.

1) Subject Id column should be named as 'Id'.

2) OS column must be named as 'OS' .

3) Death status/event column should be named as 'Death'.

4) PFS column should be named as 'PFS'.

5) Progression event column should be named as 'Prog'.
}
\examples{
SurviCompet(6,104,user_data)
}
\author{
Akash Pawar
}
